Update WANDERLAND.js
I'm not sure if I changed anything in this commit, mostly I've been struggling to fix the battle loop error. I've tried everything I could think of but nothing works. And the worst part? I know *exactly* what the problem is. But I can't fix it because I don't know how. But it's okay. This is an ongoing project and eventually, even though I don't know when, I'll figure this out.
For now though, I'll just go make assets for KittyTasker. That'll be fun. :)